1.修改ProgressView&DialProgressView的progress maxprogress为float类型 2.ProgressView增加asCircle属性，删除textInCenter属性

+++ app/src/main/java/com/benio/demoproject/progress/Formatter.java
@@ -14,5 +14,5 @@ public interface Formatter {
      * @return A formatted string representation.
      */
     @NonNull
-    public String format(int value);
+    public String format(float value);
 }
